+++ R/utilities.R
@@ -396,9 +396,11 @@ augment_newdata <- function(x, data, newdata, .se_fit, ...) {
   }
 
   resp <- safe_response(x, df)
+
   if (!is.null(resp) && is.numeric(resp)) {
-    df$.resid <- (df$.fitted - resp) %>% unname() 
+    df$.resid <- (resp - df$.fitted) %>% unname() 
   }
+
   df
 }